repo: Fixing relative vs absolute path
I am not particular fan of these, and I may change it later, but I
wanted to complain about you executing this module with `python
wwe/__main__.py` instead of setting up a console script in setup.py

I let you go because you are learning though, but you may want to
have a look into `The Python Guide to the Galaxy` and `The Python
packager guide` I think they are. They explain what is an entrypoint
and what is a console_script used for.

I think I will not do it myself to let you discover and read around

Also, I have seen you have a lot of code to test that you are
developing things properly. I would recommend you have a look on
click library to make command line clients quite seamlessly.

As a thumb rule, straight to execute code should never be in a
module, and the recommended method is to create a function and an
entrypoint associated with it. You will find however that with click
you can just setup one single command, and then handle with click
which specific functions you want to run.

There is an example on this on mosbot.